Move creation of singleton Picasso to Application onCreate This allows using cached images for notifications
