+++ app/src/main/AndroidManifest.xml
@@ -16,7 +16,8 @@
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
-        android:theme="@style/AppTheme">
+        android:theme="@style/AppTheme"
+        android:name="app.insti.InstiAppApplication">
 
         <meta-data
             android:name="com.google.android.gms.version"
